Based on the given git changes, the commit message would be:
```
feat(deps): add chrono and tracing dependencies
```

This commit message follows the Conventional Commits format and is short but descriptive of the changes made. The "feat" type indicates a new feature was added, and the "(deps)" scope specifies that the changes were to the dependencies. The message briefly describes the addition of the "chrono" and "tracing" dependencies.